% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bathy_to_ncdf.R
\name{add_bathy_to_ncdf}
\alias{add_bathy_to_ncdf}
\title{Add a matrix bathymetry to a PyGETM bathymetry netcdf file}
\usage{
add_bathy_to_ncdf(mtrx, ncdf, depth_name_new, depth_name_orig = "bathymetry")
}
\arguments{
\item{ncdf}{character; path to a PyGETM bathymetry netcdf file}

\item{var_name}{character; name of variable to extract}
}
\description{
Add a matrix bathymetry to a PyGETM bathymetry netcdf file
}
\examples{
\dontrun{
   read_bathy_as_matrix("path/to/bathy.nc", var_name = "bathymetry")
 }
}
