% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_bathy_matrix.R
\name{calc_max_diff}
\alias{calc_max_diff}
\title{Calculate maximum difference with adjacent cells}
\usage{
calc_max_diff(matrix, ind_x, ind_y)
}
\arguments{
\item{matrix}{matrix; NA not part of water, and other values are existing depths}

\item{ind_x, ind_y}{integer; indices}
}
\description{
Calculate maximum difference with adjacent cells
}
