% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_bathy_matrix.R
\name{calc_max_rx0}
\alias{calc_max_rx0}
\title{Calculate maximum rx0 between adjacent cells}
\usage{
calc_max_rx0(matrix, ind_x, ind_y)
}
\arguments{
\item{matrix}{matrix; NA not part of water, and other values are existing depths}

\item{ind_x, ind_y}{integer; indices}
}
\description{
Calculate maximum rx0 between adjacent cells
}
