% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arg_validity.R
\name{check_arg_validity}
\alias{check_arg_validity}
\title{Check the validity of the arguments to read_pygetm_output}
\usage{
check_arg_validity(ncdf, x, y, depth, z, transect, save_everything)
}
\arguments{
\item{ncdf}{character; name of the output nc file}

\item{x, y}{numeric; x and y coordinates. If NULL, extracts all.
These are the actual values, not the index.}

\item{depth}{numeric; depth below surface, should be negative. If NULL, extracts all.
These are the actual values, not the index. 'depth' and 'z' cannot be provided both.}

\item{z}{numeric; layer number (0 at bottom). If NULL, extracts all.
These are the actual values, not the index (or: index in Python-counting starting at 0).
'depth' and 'z' cannot be provided both.}

\item{transect}{data.table; headers 'x' and 'y', providing a set of cells to extract.
Can be used in combination with depth/z or not. 'If NULL, 'x' and 'y' args are
ignored of 'transect' is provided. If NULL, will not extract transect.}

\item{save_everything}{logical; if TRUE, exports entire output, without formatting. Defaults to FALSE}
}
\description{
Check the validity of the input arguments to avoid difficult-to-understand errors further on
}
