% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_interface_arg.R
\name{check_interface_arg}
\alias{check_interface_arg}
\title{Check the validity of the 'interface' argument}
\usage{
check_interface_arg(interface)
}
\arguments{
\item{interface}{list; list containing coordinates on the form list(direction, x, y),
where direction is character ("x" or "y"), and x,y numeric, representing the
x and y coordinate of the centre of the cell. Optional list arguments are "sign",
which is default +1 (positive in increasing x or y) and "interface_pos", default
+1 (the interface in the direction of direction*sign). Can be visualised with
function 'plot_interface'.}
}
\description{
Check the validity of the 'interface' argument to the flux_interface and plot_interface functions
}
