% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_regular_spacing.R
\name{check_regular_spacing}
\alias{check_regular_spacing}
\title{Check whether depth values are regularly spaced}
\usage{
check_regular_spacing(
  df,
  col_date = "date",
  col_id = col_id,
  col_depth = col_depth
)
}
\arguments{
\item{df}{data.table}

\item{col_date, col_id, col_depth}{character; column names, 'col_depth' is grouped by 'col_date' and 'col_id'.}
}
\description{
Checks if for each unique combination of 'id' and 'date', depths are regularly spaced
}
