% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_nc_time.R
\name{convert_nc_time}
\alias{convert_nc_time}
\title{Converts netcdf time to POSIXct}
\usage{
convert_nc_time(time_vals, time_att, tz = "UTC")
}
\arguments{
\item{time_vals}{numeric; returned by ncdf4::ncvar_get()}

\item{time_att}{list; returned by ncdf4::ncatt_get().}

\item{tz}{character; time zone, accepted by as.POSIXct. Defaults to UTC}
}
\description{
Converts netcdf time from numerical values to POSIXct, when providing values and attributes. The
values are the result of ncvar_get(nc, "time") and the attribute of ncatt_get(nc, "time")
}
