% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_convert.R
\name{dict_convert}
\alias{dict_convert}
\title{Translate keys into values using a dictionary}
\usage{
dict_convert(vals, dict, dict_lvl = NULL, other = as.character(NA))
}
\arguments{
\item{vals}{character; values to be compared to the names of the dictionary 
and converted.}

\item{dict}{list; values to be compared to the names of the dictionary 
and converted. All names must be unique!}

\item{dict_lvl}{character; if NULL (default), then the value is assumed to be
the value of the list. If not, the value can be one level "deeper" into the
list (see example).}

\item{other}{character; value to assign if 'vals' to not occur in the names
of the dictionary.}
}
\description{
Translate keys into values using a dictionary
}
\details{
This function works akin to Python dictionaries and a dictionary (list) with
  unique keys should be provided. This function could probably be optimised
  further. Currently only works for character values.
}
\examples{
  countries = c("France", "Netherlands", "Hungary", "Germany")
  country_code_table = list(Netherlands = list(alpha2 = "NL",
                                               alpha3 = "NLD"),
                            France = list(alpha2 = "FR",
                                          alpha3 = "FRA"),
                            Germany = list(alpha2 = "DE",
                                           alpha3 = "DEU"),
                            Hungary = list(alpha2 = "HU",
                                           alpha3 = "HUN"),
                            UK = list(alpha2 = "GB",
                                      alpha3 = "GBR"),
                            Spain = list(alpha2 = "ES",
                                         alpha3 = "ESP"))
  dict_convert(countries, country_code_table, dict_lvl = "alpha3")
}
