% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_bathy_matrix.R
\name{explore_matrix}
\alias{explore_matrix}
\title{Recursive function that checks what values in 'mat' exist at what distance.}
\usage{
explore_matrix(mat, start_i, start_j, max_scan_range)
}
\arguments{
\item{mat}{matrix; -1 are empty cells, NA not part of water, and other values are existing depths}

\item{start_i, start_j}{integer; coordinates of the reference cell}

\item{max_scan_range}{integer; the range over which the function should look for existing depths}
}
\description{
Recursive function that checks what values in 'mat' exist at what distance.
}
