% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_nc.R
\name{explore_nc}
\alias{explore_nc}
\title{List dimensions and variables in PyGETM .nc file}
\usage{
explore_nc(ncdf, var = NULL, dim = NULL)
}
\arguments{
\item{ncdf}{character; name of the output nc file}

\item{var}{character; name of a variable, default NULL (showing info about entire ncdf file if var = dim = NULL)}

\item{dim}{character; name of a dimension, default NULL (showing info about entire ncdf file if var = dim = NULL). Not used if 'var' is specified.}
}
\description{
List dimensions and variables in PyGETM .nc file
}
\details{
Show information about the PyGETM netcdf output file
}
\examples{
 \dontrun{
   explore_nc("pygetm_output.nc")
   explore_nc("pygetm_output.nc", dim = "z")
   explore_nc("pygetm_output.nc", dim = "temp")
 }
}
\author{
Jorrit Mesman
}
