% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_profile.R
\name{extract_from_profile}
\alias{extract_from_profile}
\title{Extract a value for specific depths from a profile}
\usage{
extract_from_profile(vals, depths, the_depth, depth_bott)
}
\arguments{
\item{vals}{numeric; vector of values, same length as depths}

\item{depths}{numeric; ordered vector of values, same length as depths}

\item{the_depth}{numeric; depth for which to extract the value.}

\item{depth_bott}{numeric; depth below surface of the bottom (negative).}
}
\description{
Extract a value for specific depths from a profile
}
\details{
A function to extract a value from a certain depth ("the_depth")
}
\examples{
 calc_depth_from_z(vals = c(3.5, 4, 6.8, 7.1, 8.9),
                   depths = c(-4.6, -3.9, -2.8, -1.9, -0.7),
                   the_depth = -3,
                   depth_bott = -5.2)
}
\author{
Jorrit Mesman
}
