% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_interface.R
\name{flux_interface}
\alias{flux_interface}
\title{Calculate flow from a PyGETM output file (3D)}
\usage{
flux_interface(ncdfs, interface)
}
\arguments{
\item{ncdfs}{vector; vector of character name(s) of the output nc file(s)}

\item{interface}{list; list containing coordinates on the form list(direction, x, y),
where direction is character ("x" or "y"), and x,y numeric, representing the
x and y coordinate of the centre of the cell. Optional list arguments are "sign",
which is default +1 (positive in increasing x or y) and "interface_pos", default
+1 (the interface in the direction of direction*sign). Can be visualised with
function 'plot_interface'.}
}
\description{
Calculate flow from a PyGETM output file (3D)
}
\details{
Read PyGETM 3D output files into R and calculate total flow (summed over all
  depth layers) at an interface containing several cells at each timestep.
}
\examples{
 \dontrun{
    flux_interface(ncdfs = c("output_file1.nc", "output_file2.nc"),
                   interface = list(list(x = 652650, y = 6592050, direction = "x"),
                                    list(x = 652650, y = 6591750, direction = "x")))
 }
}
\author{
Jorrit Mesman, Henrik Jonasson, Alberto Fernandez Comesana, Moutushi Sen
}
