% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dim_name.R
\name{get_dim_name}
\alias{get_dim_name}
\title{Standardise coordinates}
\usage{
get_dim_name(options, coord_names)
}
\arguments{
\item{options}{Character. Potential names to check}

\item{coord_names}{Character. Vector of provided values in netcdf file}
}
\description{
Standardise coordinates
}
\author{
Muhammed Shikhani
}
