% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_bathy_matrix.R
\name{get_dist}
\alias{get_dist}
\title{Get distance between two cells}
\usage{
get_dist(point1, point2)
}
\arguments{
\item{point1, point2}{integer; each argument is a vector '[row,col]'}
}
\description{
Get distance between two cells
}
