% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ind.R
\name{get_ind}
\alias{get_ind}
\title{Get index of closest value in a vector}
\usage{
get_ind(val, vctr, rule = "error")
}
\arguments{
\item{val}{numeric; a value}

\item{vctr}{numeric; vector of values, should be ordered}

\item{rule}{character; what to do if a vector is outside the range}
}
\description{
Get index of closest value in a vector, and throw error if the value is out of range.
In case of a tie, the function returns the first index.
}
