% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_bathy_matrix.R
\name{get_max_val}
\alias{get_max_val}
\title{Get maximum value of 'method'}
\usage{
get_max_val(matrix, method)
}
\arguments{
\item{matrix}{matrix; NA not part of water, and other values are existing depths}

\item{method}{character; 'diff' or 'rx0'}
}
\description{
Get maximum value of 'method'
}
