% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_bathy_matrix.R
\name{handle_cell}
\alias{handle_cell}
\title{Handle local smoothing for one already-filled cell}
\usage{
handle_cell(i, j)
}
\arguments{
\item{i, j}{integer; indices of the cell to smooth}
}
\description{
Handle local smoothing for one already-filled cell
}
\details{
Uses the environment of the 'smooth_bathy_matrix' function
}
