% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_bathy_matrix.R
\name{handle_step}
\alias{handle_step}
\title{Handle local smoothing step for one empty cell}
\usage{
handle_step(new_i, new_j, cell_orig)
}
\arguments{
\item{new_i, new_j}{integer; indices of the cell to smooth}

\item{cell_orig}{integer; corresponds to a cell in maintain_coords}
}
\description{
Handle local smoothing step for one empty cell
}
\details{
Uses the environment of the 'handle_cell' function
}
