% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_sim_and_obs.R
\name{pair_sim_and_obs}
\alias{pair_sim_and_obs}
\title{Pair simulations and observations}
\usage{
pair_sim_and_obs(
  ncs,
  obs,
  list_vars,
  round_depth = NULL,
  round_val = NULL,
  col_date = "date",
  col_x = "x",
  col_y = "y",
  col_depth = "depth",
  save_z = TRUE,
  max_time_diff = lubridate::days(1)
)
}
\arguments{
\item{ncs}{character; names of the output nc file}

\item{obs}{character; name of the observation file (csv format)}

\item{list_vars}{list; list to couple the simulation variables (list names) to the names of
the columns in the observation file (list values).}

\item{round_depth, round_val}{integer; Round depth and variable value to this many digits. No rounding if NULL.}

\item{col_date, col_x, col_y, col_depth}{character; column names in 'obs' file}

\item{save_z}{logical; Whether to save the z-layer from which simulation value was extracted.
Can be used to average over if you want to avoid multiple observations linking to the same z-layer.}

\item{max_time_diff}{Period; only select closest time if it is within this period
from an observation (e.g. 'days(2)'). default = 1 day. If NULL, closest point is taken. If "exact", only
exact date/time matches are used.}
}
\description{
Create tables in which GETM simulations and observations are paired. The 
observations refer to a specific time and location, which are extracted from
the netcdf outputs.
}
\examples{
 \dontrun{
   ncs = c("output_3d_20200101.nc", "output_3d_20200201.nc")
   obs = "obs_file.csv" # Columns "date", "x", "y", "depth", "temp", and "salinity"
   max_time_diff = lubridate::days(1)
   list_vars = list(temp = "temp",
                    salt = "salinity")
   
   pair_sim_and_obs(ncs = ncs, obs = obs, list_vars = list_vars)
 }
}
\author{
Jorrit Mesman, Shajar Regev
}
