% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interactive_bathymetry.R
\name{plot_interactive_bathymetry}
\alias{plot_interactive_bathymetry}
\title{Plot an interactive bathymetry plot}
\usage{
plot_interactive_bathymetry(
  ncdf,
  bathy_name = "bathymetry",
  coord_type = c("indices", "projected", "geographic")
)
}
\arguments{
\item{ncdf}{Character. Path to the NetCDF bathymetry file.}

\item{coord_type}{Character. One of 'projected', 'geographic', or 'indices'. The type of coordinates to display in the plot.}
}
\description{
Plot an interactive bathymetry plot
}
\details{
An interactive bathymetry plot to display coordinates and depth of a cell by hovering over the plot.
  By clicking, a pop-up message returns the coordinates and the depth. 
  The coordinates can be geographic ('lon', 'lat' in degrees), projected ('x', 'y' in meters), or grid indices ('i', 'j').
}
\examples{
 \dontrun{
   plot_interactive_bathymetry("Bathymetry/bathymetry.nc", coord_type = "projected")
 }
}
\author{
Muhammed Shikhani
}
