% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interface.R
\name{plot_interface}
\alias{plot_interface}
\title{Plots interface}
\usage{
plot_interface(
  ncdfs,
  interface,
  variable = "temp",
  plotting_window = 5L,
  arrow_length = 0.2,
  arrow_tip = 0.2
)
}
\arguments{
\item{ncdfs}{vector; vector of character name(s) of the output nc file(s). Only the
first argument is used for this plotting function}

\item{interface}{list; list containing coordinates on the form list(direction, x, y),
where direction is character ("x" or "y"), and x,y numeric, representing the
x and y coordinate of the centre of the cell. Optional list arguments are "sign",
which is default +1 (positive in increasing x or y) and "interface_pos", default
+1 (the interface in the direction of direction*sign).}

\item{variable}{character; function needs a variable in order to know where there
is water or not. By default 'temp'}

\item{plotting_window}{integer; only plot plotting_window cells around the area
where you defined the interface. If Inf, plots the whole bathymetry. Default = 5L}

\item{arrow_length}{numeric; length of arrow "shaft", relative to width. Might need
adjustment based on the size of your plot.}

\item{arrow_tip}{numeric; length of arrow "tip", in cm. Might need adjustment based on
the size of your plot.}
}
\description{
Plots interface
}
\details{
The primary use of this function is to provide a visual confirmation of the
  interface that is an input in the 'flux_interface' function. Therefore it uses
  the same arguments.
}
\examples{
 \dontrun{
    plot_interface(ncdfs = "output_file.nc",
                   interface = list(list(x = 652650, y = 6592050, direction = "x"),
                                    list(x = 652650, y = 6591750, direction = "x")))
 }
}
\author{
Jorrit Mesman
}
