% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sideview.R
\name{plot_sideview}
\alias{plot_sideview}
\title{Make a side-view plot}
\usage{
plot_sideview(
  df,
  var,
  col_id = "transect_id",
  col_depth = "depth",
  surface_lvl = NULL,
  na_colour = "white",
  ...
)
}
\arguments{
\item{df}{data.table; a data.table with columns that are specified in 'col_id' and 'col_depth'.}

\item{var}{character; column name in 'df', to be plotted in colour}

\item{col_id, col_depth}{character; column names of 'id' and 'depth' values}

\item{surface_lvl}{numeric; single value or vector with the level of the surface above the sediment}

\item{na_colour}{character; colour of NA values}

\item{...}{Other parameters are fed into the 'labs()' function, and 'palette' can also be specified}
}
\description{
Make a side-view plot
}
\details{
Make a side-view plot (ggplot2 object), based on a data.table with 'd' and 'depth' columns.
  If 'depth' is not there, the function will search for a 'z' column and plot that instead.
}
\examples{
 \dontrun{
   df = data.table(transect_id = c(1, 1, 1, 2, 2, 2, ...),
                   depth = c(-0.5, -1.5, -2.5, -0.5, -1.5, -2.5, ...),
                   variable = c(3.2, 1.2, 0.8, 3.6, 1.9, 0.5, ...))
   plot_topview(df, "variable", col_x = "x", col_y = "y", palette = "Blues",
               title = "Test plot")
 }
}
\author{
Jorrit Mesman
}
