% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_topview.R
\name{plot_topview}
\alias{plot_topview}
\title{Make a top-view plot}
\usage{
plot_topview(df, var, col_x = "x", col_y = "y", na_colour = "white", ...)
}
\arguments{
\item{df}{data.table; a data.table with columns that are specified in 'col_x', 'col_y', and 'var'.}

\item{var}{character; column name in 'df', to be plotted in colour}

\item{col_x, col_y}{character; column names of 'x' and 'y' coordinates}

\item{...}{Other parameters are fed into the 'labs()' function, and 'palette' can also be specified}
}
\description{
Make a top-view plot (ggplot2 object), based on a data.table with 'x' and 'y' coordinates
}
\examples{
 \dontrun{
  df = data.table(x = c(1, 2, 3, ...),
                  y = c(7, 7, 7, ...),
                  variable = c(3.2, 1.2, 0.8, ...))
  plot_topview(df, "variable", col_x = "x", col_y = "y", palette = "Blues",
               title = "Test plot")
 }
}
\author{
Jorrit Mesman
}
