% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_2d_multiple.R
\name{read_2d_multiple}
\alias{read_2d_multiple}
\title{Read multiple PyGETM 2D output files and merge them}
\usage{
read_2d_multiple(ncdfs, var, round_val = NULL)
}
\arguments{
\item{ncdfs}{character; vector with the names of the output nc files}

\item{var}{character; name of the variable in the nc file. Can be a vector.}

\item{round_val}{integer; Round depth and variable value to this many digits. No rounding if NULL.}
}
\description{
Read multiple PyGETM 2D output files and merge them
}
\details{
Repeated calls to 'read_pygetm_output_2d' and merging them into a single
  data.table. A common way of running PyGETM is to save outputs from different
  months into separate netcdf files, and this function would merge them into
  one data.table.
}
\examples{
 \dontrun{
 read_2d_multiple(ncdfs = c("output2d_20210101.nc", "output2d_20210201.nc"),
                    var = "zt",
                    round_val = 5L)
 }
}
\author{
Jorrit Mesman
}
