% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bathy_as_matrix.R
\name{read_bathy_as_matrix}
\alias{read_bathy_as_matrix}
\title{Read bathymetry as a matrix}
\usage{
read_bathy_as_matrix(ncdf, var_name = "bathymetry")
}
\arguments{
\item{ncdf}{character; path to a PyGETM bathymetry netcdf file}

\item{var_name}{character; name of variable to extract}
}
\description{
Read bathymetry as a matrix
}
\details{
This function can also be used to read another variable from the bathymetry
ncdf file (e.g. x and y coordinates)
}
\examples{
\dontrun{
   read_bathy_as_matrix("path/to/bathy.nc", var_name = "bathymetry")
 }
}
