% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_coord_maintain_file.R
\name{read_coord_maintain_file}
\alias{read_coord_maintain_file}
\title{Read x and y coordinates from a csv file into indices and return data.table}
\usage{
read_coord_maintain_file(filepath, ncdf)
}
\arguments{
\item{filepath}{character; path to a csv file with columns "x" and "y"
(or "ind_x" and "ind_y" already). Optional column 'fix_depth'; any values
filled here are used instead of the original bathy's depth.}

\item{ncdf}{character; path to a PyGETM bathymetry netcdf file}
}
\description{
Read x and y coordinates from a csv file into indices and return data.table
}
\examples{
\dontrun{
   read_coord_maintain_file("path/to/coord_maintain.csv", "path/to/bathy.nc")
 }
}
