% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_local_multiple.R
\name{read_local_multiple}
\alias{read_local_multiple}
\title{Read multiple local PyGETM output files and merge them}
\usage{
read_local_multiple(
  ncdfs,
  var,
  depth = NULL,
  z = NULL,
  round_depth = NULL,
  round_val = NULL,
  profile_interval = NULL
)
}
\arguments{
\item{ncdfs}{character; vector with the names of the output nc files}

\item{var}{character; name of the variable in the nc file}

\item{depth}{numeric; depth below surface, should be negative. If NULL, extracts all.
'depth' and 'z' cannot be provided both.}

\item{z}{numeric; layer number (0 at bottom). If NULL, extracts all.
These are the actual values, not the index (or: index in Python-counting starting at 0).
'depth' and 'z' cannot be provided both.}

\item{round_depth, round_val}{integer; Round depth and variable value to this many digits. No rounding if NULL.}

\item{profile_interval}{numeric; single value, calculates output depths for a profile with
this depth interval}
}
\description{
Read multiple local PyGETM output files and merge them
}
\details{
Repeated calls to 'read_local' and merging them into a single
  data.table. A common way of running PyGETM is to save outputs from different
  months into separate netcdf files, and this function would merge them into
  one data.table.
  
  There are Several options for extracting output.
}
\examples{
 \dontrun{
 read_multiple(ncdfs = c("output_local_20210101.nc", "output_local_20210201.nc"),
                    var = "temp",
                    depth = NULL,
                    z = 19,
                    round_depth = 2L,
                    round_val = 3L)
 }
}
\author{
Jorrit Mesman
}
