% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pygetm_output_2d.R
\name{read_pygetm_output_2d}
\alias{read_pygetm_output_2d}
\title{Read a PyGETM output file (2D)}
\usage{
read_pygetm_output_2d(ncdf, var, round_val = NULL)
}
\arguments{
\item{ncdf}{character; name of the output nc file}

\item{var}{character; name of the variable in the nc file. Can be a vector.}

\item{round_val}{integer; Round depth and variable value to this many digits. No rounding if NULL.}
}
\description{
Read a PyGETM output file (2D)
}
\details{
Read a PyGETM 2D output file into R.
}
\examples{
 \dontrun{
  read_pygetm_output_2d(ncdf = "pygetm_output_2d.nc",
                        var = "zt",
                        round_val = 3L)
 }
}
\author{
Jorrit Mesman
}
