% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_bathy_matrix.R
\name{rx0}
\alias{rx0}
\title{Calculate rx0}
\usage{
rx0(h1, h2)
}
\arguments{
\item{h1, h2}{numeric; depths}
}
\description{
Calculate rx0
}
\details{
delta-H / (2 * mean(H), Beckmann & Haidvogel 1993, doi:10.1175/1520-0485(1993)023<1736:NSOFAA>2.0.CO;2
}
