% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_bathy_matrix.R
\name{rx0_inv}
\alias{rx0_inv}
\title{Calculate depth of h2, if you know h1 and rx0}
\usage{
rx0_inv(h1, rx0, sign = 1)
}
\arguments{
\item{h1}{numeric; depth of a cell}

\item{rx0}{numeric; slope parameter}

\item{sign}{numeric; if positive, increase in depth, if negative, decrease}
}
\description{
Calculate depth of h2, if you know h1 and rx0
}
