% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_matrix_local.R
\name{slice_matrix_local}
\alias{slice_matrix_local}
\title{Extract part of a matrix}
\usage{
slice_matrix_local(
  lst_mtrx,
  depth,
  z,
  mtrx_zct = NULL,
  mtrx_surf = NULL,
  mtrx_bott = NULL,
  add_depth_to_output = T
)
}
\arguments{
\item{lst_mtrx}{list; list of matrices of the variables to extract. Result of ncdf4::ncvar_get(...)}

\item{depth}{numeric; depth below surface, should be negative. If NULL, extracts all.
'depth' and 'z' cannot be provided both.}

\item{z}{numeric; layer number (1 at bottom). If NULL, extracts all.
Note: NOT in Python-counting (so first one is "1"). 'depth' and 'z' cannot be provided both.}

\item{mtrx_zct}{matrix: matrix of 'zct'. Result of ncdf4::ncvar_get(...)}

\item{mtrx_surf, mtrx_bott}{matrix; similar to mtrx_zct, but based on a matrix
of the interfaces ('zft') and given for the uppermost and lowermost layer, respectively}

\item{add_depth_to_output}{logical; if true and 'depth' is not provided, then
still 'depth' is calculated for each value of 'z'}
}
\description{
Extract part of a matrix
}
\details{
Takes a list of matrices (as gotten from ncvar_get() on a PyGETM local output file) and
  extracts part of them, based on the input arguments. This function is for internal use in
  read_local.R.
}
\author{
Jorrit Mesman
}
