% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_matrix_with_obs.R
\name{slice_matrix_with_obs}
\alias{slice_matrix_with_obs}
\title{Extract part of a matrix, based on observations}
\usage{
slice_matrix_with_obs(
  mtrx,
  mtrx_zft,
  x_dim,
  y_dim,
  time_dim,
  observations,
  col_x = "x",
  col_y = "y",
  sim_date = "sim_date",
  col_depth = "depth",
  save_z = TRUE
)
}
\arguments{
\item{mtrx}{matrix; matrix of the variable to extract. Result of ncdf4::ncvar_get(...)}

\item{mtrx_zft}{matrix: matrix of 'zft'. Result of ncdf4::ncvar_get(...)}

\item{x_dim, y_dim}{numeric; vector of the values of the 'x' and 'y' dimensions}

\item{time_dim}{POSIXct; vector of the values of the 'time' dimension in the nc file}

\item{col_x, col_y, col_depth}{character; column names in 'observations' for the dimensions}

\item{sim_date}{character; column name of the simulation date}

\item{save_z}{logical; Whether to save the z-layer from which simulation value was extracted.
Can be used to average over if you want to avoid multiple observations linking to the same z-layer.}
}
\description{
Extract part of a matrix, based on observations
}
\details{
Loops over a set of provided observations and extracts values for the
  closest locations and times. Only the closest cells are extracted, no interpolation
  is done. This function is for internal use in pair_sim_and_obs.R.
}
\author{
Jorrit Mesman
}
