% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_bathy_matrix.R
\name{smooth_bathy_matrix}
\alias{smooth_bathy_matrix}
\title{Function to smooth a bathymetry matrix}
\usage{
smooth_bathy_matrix(
  mtrx,
  maintain_coords,
  max_val,
  method = "diff",
  track_volume = TRUE,
  min_depth = 0,
  global_smoothing = TRUE,
  box_width_global = 4L,
  box_width_shallows = 6L,
  adjust_factor_global = 0.05,
  max_iter_global = 10L,
  inflexibility = NULL,
  max_scan_range = max(5L, length(inflexibility) * 2),
  max_vol_adj_step_local = 0.5,
  bathy_round = Inf,
  warn = T,
  margin = 1e-05
)
}
\arguments{
\item{mtrx}{matrix; numeric matrix, describing a bathymetry, Can be read from a netcdf
file by 'read_bathy_as_matrix'.}

\item{maintain_coords}{data.table; should have columns 'ind_x' and 'ind_y', which
should correspond to values in 'mtrx' that should not be affected by the smoothing.
If NULL, no points are maintained and only 'global' smoothing is applied. Can be
read from a csv with "x" and "y" coordinates using 'read_coord_maintain_file'}

\item{max_val}{numeric; maximum allowed value of 'method'. Should always be positive.}

\item{method}{character; 'diff' or 'rx0'. If 'diff', 'max_val' is interpreted as
the maximum (absolute) difference between adjacent cells. If 'rx0', 'max_val' is
the maximum slope parameter 'rx0', following Beckmann & Haidvogel (1993)}

\item{track_volume}{logical; if TRUE, the algorithm strives to maintain the original
volume of the provided bathymetry.}

\item{min_depth}{numeric; minimum allowed depth (always positive). Defaults to 0.0}

\item{global_smoothing}{logical; if TRUE, global smoothing is applied}

\item{box_width_global}{integer; width of box used in global smoothing}

\item{box_width_shallows}{integer; width of box used in global smoothing when assessing
'min_depth'}

\item{adjust_factor_global}{numeric; [0-1] during global smoothing, in each step, the box
is adjusted by this much towards the mean of the box. If 0, no adjustment happens, if 1,
the whole box is set to the mean.}

\item{max_iter_global}{integer; maximum amount of iteration in global smoothing}

\item{inflexibility}{numeric; vector of values between 1 and 0. Used during local smoothing.
Describes the extent to which values around maintain_coords are smoothed. If you want to
keep the areas around maintain_coords similar to that depth, provide more (high) numbers.
Example: c(0.9, 0.7, 0.5, 0.3, 0.1). If NULL, no inflexibility is used.}

\item{max_scan_range}{integer; used during local smoothing. Longer range can prevent
impossible constraints to occur, but a too long range can slow down the function.}

\item{max_vol_adj_step_local}{numeric; maximum allowed adjustment to the depth in order
to correct the volume imbalance. Primary parameter to change if volume imbalance exists
after smoothing. Not used if track_volume is FALSE.}

\item{bathy_round}{integer; optional parameter to round the bathymetry to this many digits}

\item{warn}{logical; throw a warning if some maintain_coords are shallower than 'min_depth'}

\item{margin}{numeric; low number to prevent errors in calculating volume imbalances and constraints}
}
\description{
Function to smooth a bathymetry matrix
}
\details{
Takes a matrix and applies smoothing algorithms, including options to
  maintain the depth of certain designated points.
  
  Two phases are used in the function: 'global' smoothing where a window is
  moved over the matrix and steep slopes are smoothed and (optionally) a
  minimum depth is set, without maintaining the designated points. Then a
  'local' smoothing is done starting from each designated point and the
  globally-smoothed bathymetry is adjusted to maintain the depths of the 
  points and still fulfill the smoothing criteria. Either step is optional.
  
  Smoothing is foremost an iterative process. Start with a certain parameter
  set and plot the results using 'plot_bathy' and then adjust. The common workflow
  is: 'read_bathy_as_matrix', (iteratively) 'smooth_bathy_matrix' & 'plot_bathy', then
  'add_bathy_to_ncdf'.
}
\examples{
mat = matrix(c(NA, 3.1, 1.1, 10.1, NA,
               NA, NA, 14.5, 19.2, NA,
               NA, 1.0, 7.7, 22.0, 16.5,
               4.2, 3.5, 6.9, 13.0, 12.7,
               NA, NA, 5.1, NA, NA),
             nrow = 5L, ncol = 5L)
df_maintain = data.table(ind_x = 4, ind_y = 2) # mat[4,2] == 19.2
# Smooth using 'diff' and no inflexibility
smooth_bathy_matrix(mtrx = mat, maintain_coords = df_maintain,
                    max_val = 6, method = "diff",
                    max_vol_adj_step_local = 2.5)
# Smooth using 'rx0', min_depth, and inflexibility
smooth_bathy_matrix(mtrx = mat, maintain_coords = df_maintain,
                    max_val = 0.25, method = "rx0",
                    min_depth = 5.0,
                    inflexibility = c(0.5, 0.2),
                    max_vol_adj_step_local = 6.0)
 
}
\author{
Jorrit Mesman
}
