% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_bathy_matrix.R
\name{smooth_shallow_area}
\alias{smooth_shallow_area}
\title{Global smoothing function for shallow areas}
\usage{
smooth_shallow_area(
  bathy,
  min_depth,
  max_iter = 20L,
  box_width = 4L,
  adjust_factor = 0.05
)
}
\arguments{
\item{bathy}{matrix; NA not part of water, and other values are existing depths}

\item{min_depth}{numeric; minimum allowed depth}

\item{max_iter}{integer; maximum number of iterations}

\item{box_width}{integer; width of moving window}

\item{adjust_factor}{numeric; between 0 (no adjustment) and 1 (setting whole box
to the mean)}
}
\description{
Global smoothing function for shallow areas
}
