% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_bathy_matrix.R
\name{smooth_steep_slopes}
\alias{smooth_steep_slopes}
\title{Global smoothing function}
\usage{
smooth_steep_slopes(
  bathy,
  method,
  max_val,
  max_iter = 20L,
  box_width = 4L,
  adjust_factor = 0.05
)
}
\arguments{
\item{bathy}{matrix; NA not part of water, and other values are existing depths}

\item{method}{character; 'diff' or 'rx0'}

\item{max_val}{numeric; maximum allowed value of 'method'}

\item{max_iter}{integer; maximum number of iterations}

\item{box_width}{integer; width of moving window}

\item{adjust_factor}{numeric; between 0 (no adjustment) and 1 (setting whole box
to the mean)}
}
\description{
Global smoothing function
}
