% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_transect.R
\name{visualise_transect}
\alias{visualise_transect}
\alias{visualize_transect}
\title{Plot a transect}
\usage{
visualise_transect(transect, bathy_nc, single_point_colour = TRUE, ...)

visualize_transect(transect, bathy_nc, single_point_colour = TRUE, ...)
}
\arguments{
\item{transect}{data.table; a data.table with columns 'x', and 'y'.}

\item{bathy_nc}{character; name of the bathymetry nc file that is used as input into PyGETM}

\item{...}{Other parameters are fed into the 'labs()' function, and 'palette' can also be specified}
}
\description{
Plot a transect on top of a map of the bathymetry
}
\examples{
 \dontrun{
  df = data.table(transect_id = c(1,2,3, ...),
                  x = c(200, 200, 200, ...),
                  y = c(300, 350, 400, ...))
  visualise_transect(transect = df, "bathymetry.nc")
 }
}
\author{
Jorrit Mesman
}
