% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bathy.R
\name{plot_bathy}
\alias{plot_bathy}
\title{Plot a bathymetry}
\usage{
plot_bathy(mtrx, mtrx_ref = NULL, rev_x = FALSE, rev_y = FALSE)
}
\arguments{
\item{mtrx}{matrix; numeric matrix, describing a bathymetry}

\item{mtrx_ref}{matrix; numeric matrix, describing a bathymetry. If provided,
plot will show the difference between mtrx - mtrx_ref. Default is NULL.}

\item{rev_x, rev_y}{logical; if TRUE, will plot that axis in reverse}
}
\description{
Plot a bathymetry
}
\details{
Creates a ggplot2 object that can be modified further.
  
  Takes a matrix. Can be gotten from a netcdf file by 'read_bathy_as_matrix'
  
  Either plots the bathymetry or, if 'bathy_ref' is provided, the difference
  between the two bathymetries.
  
  There is another version of this function - 'plot_interactive_bathymetry' - where
  you can check individual cell values in the RStudio viewer. That only works when
  providing a netcdf file
}
\examples{
mat = matrix(c(NA, 3.1, 1.1, 10.1, NA,
               NA, NA, 14.5, 19.2, NA,
               NA, 1.0, 7.7, 22.0, 16.5,
               4.2, 3.5, 6.9, 13.0, 12.7,
               NA, NA, 5.1, NA, NA),
             nrow = 5L, ncol = 5L)
plot_bathy(mat)
 
}
\author{
Jorrit Mesman
}
