% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_local.R
\name{read_local}
\alias{read_local}
\title{Read a PyGETM local output file (1D)}
\usage{
read_local(
  ncdf,
  var,
  depth = NULL,
  z = NULL,
  round_depth = NULL,
  round_val = NULL,
  profile_interval = NULL
)
}
\arguments{
\item{ncdf}{character; name of the local output nc file}

\item{var}{character; name of the variable in the nc file. Can be a vector.}

\item{depth}{numeric; depth below surface, should be negative. Can be a vector.  If NULL, extracts all.
'depth' and 'z' cannot be provided both.}

\item{z}{numeric; layer number (0 at bottom). If NULL, extracts all.
These are the actual values, not the index (or: index in Python-counting starting at 0).
'depth' and 'z' cannot be provided both.}

\item{round_depth, round_val}{integer; Round depth and variable value to this many digits. No rounding if NULL.
'round_val' can be a single value or have same length as 'var'.}

\item{profile_interval}{numeric; single value, calculates output depths for a profile with
this depth interval. Overwrites depth.}
}
\description{
Read a PyGETM local output file (1D)
}
\details{
Read a PyGETM local output file into R, with options for extracting certain cells.
  Option to extract a specific z or depth, or a regular interpolated interval.
  pyGETM local output is created by the "transforms" argument in output.request
  and "pygetm.output.operators.IndexXY.parameterize"
}
\examples{
 \dontrun{
 read_local(ncdf = "local_output.nc",
                    var = "temp",
                    depth = NULL,
                    z = NULL,
                    round_depth = 2L,
                    round_val = 3L)
 }
}
\author{
Jorrit Mesman
}
